I ran out of time since so here is my actual final
If possible can you grade this I ran out of time as I spent over 1 hour and 30 minutes on the previous 1 just writing comments as I am a slow typer so I had less than 1 hour to finish this section. In this version I just fixed the print statements and ironed out 1 bug with the random section.